Add explore functions
These functions are modified versions of the functions os.listall, os.listfiles and os.listdirs in the Into The Breach mod loader. They all will accept both a fixed path starting with a drive letter (i.e. C:\), as well as relative paths from the Into The Breach install directory when not starting with a drive letter.